Update betterflat.py
Included check if dsf coordinates are valid for coordinates given by airport boundary.

Also changing height of vertices that are on same coordinates as adapted triangle vertices. This is necessary, as the triangle can have vertices outside boundary, where other vertices are on same coordinates and would therefore not be changed.